+++ tests/browser/wait_for_request.test.js
@@ -2,6 +2,8 @@
 
 const Wendigo = require('../../lib/wendigo');
 const configUrls = require('../config.json').urls;
+const utils = require('../test_utils');
+
 
 describe("Wait For Request", function() {
     this.timeout(5000);
@@ -50,5 +52,17 @@ describe("Wait For Request", function() {
         await browser.assert.request.url(/api/).responseBody("test").exactly(0);
         await browser.waitForResponse("http://localhost:3456/api");
         await browser.assert.request.url(/api/).responseBody("test");
+    });
+
+    it("Wait For Request Timeout", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.waitForRequest("http://localhost:3456/api", 10);
+        }, `TimeoutError: Waiting for request "http://localhost:3456/api", timeout of 10ms exceeded.`);
+    });
+
+    it("Wait For Response Timeout", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.waitForResponse("http://localhost:3456/api", 10);
+        }, `TimeoutError: Waiting for response "http://localhost:3456/api", timeout of 10ms exceeded.`);
     });
 });