+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+1.2.0 / 2018-09-29
+==================
+
+  * Support for index in clickText
+  * Now TimeoutError will be thrown on timeouts
+  * Changed default viewport res to 1440x9000
+  * Browser.frames
+  * Puppeteer updated to 1.8
+  * Error handling improved
+
 1.1.0 / 2018-09-05
 ==================